Complete support for Passwordless tests
Closes #9850
