+++ testsuite/integration-arquillian/tests/other/webauthn/src/main/java/org/keycloak/testsuite/webauthn/utils/WebAuthnRealmData.java
@@ -20,6 +20,7 @@ package org.keycloak.testsuite.webauthn.utils;
 import org.keycloak.representations.idm.RealmRepresentation;
 
 import java.util.List;
+import java.util.function.Consumer;
 
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -78,5 +79,85 @@ public class WebAuthnRealmData {
 
     public List<String> getAcceptableAaguids() {
         return isPasswordless ? realm.getWebAuthnPolicyPasswordlessAcceptableAaguids() : realm.getWebAuthnPolicyAcceptableAaguids();
+    }
+
+    public RealmRepresentation getRealm() {
+        return realm;
+    }
+
+    public Builder builder() {
+        return new Builder(realm, isPasswordless);
+    }
+
+    public static class Builder {
+        private final RealmRepresentation realm;
+        private final boolean isPasswordless;
+
+        public Builder(RealmRepresentation realm, boolean isPasswordless) {
+            this.realm = realm;
+            this.isPasswordless = isPasswordless;
+        }
+
+        public Builder rpEntityName(String entityName) {
+            setProperty(entityName, realm::setWebAuthnPolicyRpEntityName, realm::setWebAuthnPolicyPasswordlessRpEntityName);
+            return this;
+        }
+
+        public Builder signatureAlgorithms(List<String> list) {
+            setProperty(list, realm::setWebAuthnPolicySignatureAlgorithms, realm::setWebAuthnPolicyPasswordlessSignatureAlgorithms);
+            return this;
+        }
+
+        public Builder rpId(String rpId) {
+            setProperty(rpId, realm::setWebAuthnPolicyRpId, realm::setWebAuthnPolicyPasswordlessRpId);
+            return this;
+        }
+
+        public Builder attestationConveyancePreference(String preference) {
+            setProperty(preference, realm::setWebAuthnPolicyAttestationConveyancePreference, realm::setWebAuthnPolicyPasswordlessAttestationConveyancePreference);
+            return this;
+        }
+
+        public Builder authenticatorAttachment(String attachment) {
+            setProperty(attachment, realm::setWebAuthnPolicyAuthenticatorAttachment, realm::setWebAuthnPolicyPasswordlessAuthenticatorAttachment);
+            return this;
+        }
+
+        public Builder requireResidentKey(String requirement) {
+            setProperty(requirement, realm::setWebAuthnPolicyRequireResidentKey, realm::setWebAuthnPolicyPasswordlessRequireResidentKey);
+            return this;
+        }
+
+        public Builder userVerificationRequirement(String requirement) {
+            setProperty(requirement, realm::setWebAuthnPolicyUserVerificationRequirement, realm::setWebAuthnPolicyPasswordlessUserVerificationRequirement);
+            return this;
+        }
+
+        public Builder timeout(Integer timeout) {
+            setProperty(timeout, realm::setWebAuthnPolicyCreateTimeout, realm::setWebAuthnPolicyPasswordlessCreateTimeout);
+            return this;
+        }
+
+        public Builder avoidSameAuthenticatorRegister(Boolean state) {
+            setProperty(state, realm::setWebAuthnPolicyAvoidSameAuthenticatorRegister, realm::setWebAuthnPolicyPasswordlessAvoidSameAuthenticatorRegister);
+            return this;
+        }
+
+        public Builder acceptableAaguids(List<String> aaguids) {
+            setProperty(aaguids, realm::setWebAuthnPolicyAcceptableAaguids, realm::setWebAuthnPolicyPasswordlessAcceptableAaguids);
+            return this;
+        }
+
+        public RealmRepresentation build() {
+            return realm;
+        }
+
+        private <T> void setProperty(T value, Consumer<T> webauthnSetter, Consumer<T> passwordlessSetter) {
+            if (isPasswordless) {
+                passwordlessSetter.accept(value);
+            } else {
+                webauthnSetter.accept(value);
+            }
+        }
     }
 }
