+++ testsuite/integration-arquillian/tests/other/webauthn/src/test/java/org/keycloak/testsuite/webauthn/passwordless/WebAuthnPwdLessPropertyTest.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.keycloak.testsuite.webauthn.passwordless;
+
+import org.junit.Ignore;
+import org.junit.Test;
+import org.keycloak.testsuite.webauthn.WebAuthnPropertyTest;
+
+/**
+ * @author <a href="mailto:mabartos@redhat.com">Martin Bartos</a>
+ */
+public class WebAuthnPwdLessPropertyTest extends WebAuthnPropertyTest {
+
+    @Override
+    public boolean isPasswordless() {
+        return true;
+    }
+
+    @Override
+    @Ignore("Not usable for Passwordless")
+    @Test
+    public void requiredActionRegistration() {
+    }
+}
